Hook up help link even for AbstractCodeQualityDiagnosticAnalyzer
